+++ test/retry_exhausted_test.rb
@@ -187,7 +187,7 @@ describe "sidekiq_retries_exhausted" do
     assert_equal 0, Sidekiq::DeadSet.new.size
   end
 
-  it "does not allow disabling global failure handlers when disabling dead set" do
+  it "does not disable death handlers when disabling dead set" do
     exhausted_job = nil
     exhausted_exception = nil
     @config.death_handlers.clear
@@ -205,7 +205,7 @@ describe "sidekiq_retries_exhausted" do
     assert exhausted_exception
   end
 
-  it "supports discard option to disable global failure handlers and dead set" do
+  it ":discard still calls death handlers" do
     discard_job = DiscardJob.new
 
     exhausted_job = nil
@@ -223,8 +223,8 @@ describe "sidekiq_retries_exhausted" do
 
     assert DiscardJob.exhausted_called
     assert_equal 0, Sidekiq::DeadSet.new.size
-    assert_nil exhausted_job
-    assert_nil exhausted_exception
+    refute_nil exhausted_job
+    refute_nil exhausted_exception
   end
 
   it "supports wrapped jobs" do